requirements.txt
python-rtmidi>=1.5.8
mido>=1.3.2
numpy>=1.26.0
pyfluidsynth>=1.3.3
